dashboardPage {shinydashboard}	R Documentation

Dashboard page

Description

This creates a dashboard page for use in a Shiny app.

Usage

dashboardPage(header, sidebar, body, title = NULL, skin = c("blue", "black",
  "purple", "green", "red", "yellow"))

Arguments

header
A header created by dashboardHeader.

sidebar
A sidebar created by dashboardSidebar.

body
A body created by dashboardBody.

title
A title to display in the browser's title bar. If no value is provided, it will try to extract the title from the dashboardHeader.

skin
A color theme. One of "blue", "black", "purple", "green", "red", or "yellow".

See Also

dashboardHeader, dashboardSidebar, dashboardBody.

Examples

## Only run this example in interactive R sessions
if (interactive()) {
# Basic dashboard page template
library(shiny)
shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(),
    title = "Dashboard example"
  ),
  server = function(input, output) { }
)
}

[Package shinydashboard version 0.7.1 Index]