infoBox {shinydashboard}	R Documentation

Create an info box for the main body of a dashboard.

Description

An info box displays a large icon on the left side, and a title, value (usually a number), and an optional smaller subtitle on the right side. Info boxes are meant to be placed in the main body of a dashboard.

Usage

infoBox(title, value = NULL, subtitle = NULL,
  icon = shiny::icon("bar-chart"), color = "aqua", width = 4,
  href = NULL, fill = FALSE)

Arguments

title
Title text.

value
The value to display in the box. Usually a number or short text.

subtitle
Subtitle text (optional).

icon
An icon tag, created by icon.

color
A color for the box. Valid colors are listed in validColors.

width
The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

href
An optional URL to link to.

fill
If FALSE (the default), use a white background for the content, and the color argument for the background of the icon. If TRUE, use the color argument for the background of the content; the icon will use the same color with a slightly darkened background.

See Also

box for usage examples.

Other boxes: box, tabBox, valueBox

[Package shinydashboard version 0.7.1 Index]