messageItem {shinydashboard}	R Documentation

Create a message item to place in a dropdown message menu

Description

Create a message item to place in a dropdown message menu

Usage

messageItem(from, message, icon = shiny::icon("user"), time = NULL,
  href = NULL)

Arguments

from
Who the message is from.

message
Text of the message.

icon
An icon tag, created by icon.

time
String representing the time the message was sent. Any string may be used. For example, it could be a relative date/time like "5 minutes", "today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45". If NULL, no time will be displayed.

href
An optional URL to link to.

See Also

dashboardHeader for example usage.

Other menu items: notificationItem, taskItem

[Package shinydashboard version 0.7.1 Index]