notificationItem {shinydashboard}	R Documentation

Create a notification item to place in a dropdown notification menu

Description

Create a notification item to place in a dropdown notification menu

Usage

notificationItem(text, icon = shiny::icon("warning"), status = "success",
  href = NULL)

Arguments

text
The notification text.

icon
An icon tag, created by icon.

status
The status of the item This determines the item's background color. Valid statuses are listed in validStatuses.

href
An optional URL to link to.

See Also

dashboardHeader for example usage.

Other menu items: messageItem, taskItem

[Package shinydashboard version 0.7.1 Index]