sidebarSearchForm {shinydashboard}	R Documentation

Create a search form to place in a sidebar

Description

A search form consists of a text input field and a search button.

Usage

sidebarSearchForm(textId, buttonId, label = "Search...",
  icon = shiny::icon("search"))

Arguments

textId
Shiny input ID for the text input box.

buttonId
Shiny input ID for the search button (which functions like an actionButton).

label
Text label to display inside the search box.

icon
An icon tag, created by icon.

See Also

dashboardSidebar for example usage.

Other sidebar items: sidebarMenu, sidebarUserPanel

[Package shinydashboard version 0.7.1 Index]