sidebarUserPanel {shinydashboard}	R Documentation

A panel displaying user information in a sidebar

Description

A panel displaying user information in a sidebar

Usage

sidebarUserPanel(name, subtitle = NULL, image = NULL)

Arguments

name
Name of the user.

subtitle
Text or HTML to be shown below the name.

image
A filename or URL to use for an image of the person. If it is a local file, the image should be contained under the www/ subdirectory of the application.

See Also

dashboardSidebar for example usage.

Other sidebar items: sidebarMenu, sidebarSearchForm

[Package shinydashboard version 0.7.1 Index]