tabItem {shinydashboard}	R Documentation

One tab to put inside a tab items container

Description

One tab to put inside a tab items container

Usage

tabItem(tabName = NULL, ...)

Arguments

tabName
The name of a tab. This must correspond to the tabName of a menuItem or menuSubItem.

...
Contents of the tab.

See Also

menuItem, menuSubItem, tabItems. See sidebarMenu for a usage example.

[Package shinydashboard version 0.7.1 Index]