taskItem {shinydashboard}	R Documentation

Create a task item to place in a dropdown task menu

Description

Create a task item to place in a dropdown task menu

Usage

taskItem(text, value = 0, color = "aqua", href = NULL)

Arguments

text
The task text.

value
A percent value to use for the bar.

color
A color for the bar. Valid colors are listed in validColors.

href
An optional URL to link to.

See Also

dashboardHeader for example usage.

Other menu items: messageItem, notificationItem

[Package shinydashboard version 0.7.1 Index]