validStatuses {shinydashboard}	R Documentation

Valid statuses

Description

These status strings correspond to colors as defined in Bootstrap's CSS. Although the colors can vary depending on the particular CSS selector, they generally appear as follows:

Details

primary Blue (sometimes dark blue)

success Green

info Blue

warning Orange

danger Red

[Package shinydashboard version 0.7.1 Index]