valueBox {shinydashboard}	R Documentation

Create a value box for the main body of a dashboard.

Description

A value box displays a value (usually a number) in large text, with a smaller subtitle beneath, and a large icon on the right side. Value boxes are meant to be placed in the main body of a dashboard.

Usage

valueBox(value, subtitle, icon = NULL, color = "aqua", width = 4,
  href = NULL)

Arguments

value
The value to display in the box. Usually a number or short text.

subtitle
Subtitle text.

icon
An icon tag, created by icon.

color
A color for the box. Valid colors are listed in validColors.

width
The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

href
An optional URL to link to.

See Also

box for usage examples.

Other boxes: box, infoBox, tabBox

[Package shinydashboard version 0.7.1 Index]