valueBoxOutput {shinydashboard}	R Documentation

Create an info or value box output (client side)

Description

This is the UI-side function for creating a dynamic valueBox or infoBox.

Usage

valueBoxOutput(outputId, width = 4)

infoBoxOutput(outputId, width = 4)

Arguments

outputId
Output variable name.

width
The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

See Also

renderValueBox for the corresponding server-side function and examples.

[Package shinydashboard version 0.7.1 Index]