menuOutput {shinydashboard}	R Documentation

Create a dynamic menu output for shinydashboard (client side)

Description

This can be used as a placeholder for dynamically-generated dropdownMenu, notificationItem, messageItem, taskItem sidebarMenu, or menuItem. If called directly, you must make sure to supply the correct type of tag. It is simpler to use the wrapper functions if present; for example, dropdownMenuOutput and sidebarMenuOutput.

Usage

menuOutput(outputId, tag = tags$li)

Arguments

outputId
Output variable name.

tag
A tag function, like tags$li or tags$ul.

See Also

renderMenu for the corresponding server side function and examples.

Other menu outputs: dropdownMenuOutput, menuItemOutput, renderMenu, sidebarMenuOutput

[Package shinydashboard version 0.7.1 Index]